fix: return fn symbol from generated body
This commit changes the `generate-defn` return value to be the symbol of
the newly generated function name. This is inline with the behavior of
`clojure.core/defn` and `clojure.core/defn-`.

The reason this is useful is because some tools (namely cider / nrepl) rely on
the return value of the `defn` for certain functionality, such as
debugging and tracing. By returning the symbol, those tools go back to
working with ghostwheel generated functions.